Fix compile error with latest libc++ on MacOS
static_assert(__libcpp_random_is_valid_urng<_URNG>::value, "");

+++ test/catch.hpp
@@ -7118,7 +7118,7 @@ namespace Catch {
 namespace Catch {
 
     struct RandomNumberGenerator {
-        typedef std::ptrdiff_t result_type;
+        typedef std::uint64_t result_type;
 
         result_type operator()( result_type n ) const { return std::rand() % n; }
